fix(adopt): inspect argv for tmux Cursor agent discovery
Tmux adoption previously checked only /proc/<pid>/comm when walking a pane's process tree. Cursor Agent can report comm as MainThread while the real executable lives in argv[0], so Cursor-backed /adopt still missed locally started agent processes.

Move comm+argv CLI detection into the shared session discovery module and reuse it from zellij discovery to keep both multiplexer paths consistent.
